agents/web_crawler/requirements.txt
crawl4ai>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
playwright>=1.41.0  # Required by crawl4ai for browser automation
psycopg2-binary>=2.9.9
psutil>=5.9.0  # For memory usage monitoring
fastapi>=0.110.0
uvicorn>=0.27.0 